Add Code Quality GitHub workflow (#2)
# Pull Request

## Description

This change introduces a new GitHub Actions workflow for code quality
checks. The workflow, named "Code Quality", is triggered on pushes to
the main branch and pull request events.

The workflow uses the Super-Linter tool to perform linting and
formatting on the codebase. It runs on Ubuntu and includes the following
key features:

- Checks the entire codebase on each run
- Uses a slim version of Super-Linter (v7.1.0)
- Validates YAML files with strict error handling
- Utilises custom linter configurations stored in
`.github/super-linter-configurations`
- Requires specific permissions for reading contents and packages, and
writing statuses

This addition will help maintain consistent code quality across the
project by automatically checking for style and formatting issues in the
codebase.