Fix Dependency Review (#22)
# Pull Request

## Description

This pull request includes updates to the GitHub workflows and the pull
request template. The most important changes involve the removal and
addition of the dependency review step in different workflow files.

Changes to GitHub workflows:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L51-L60):
Removed the `dependency-review` job.
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR36-R45):
Added the `dependency-review` job.

Changes to pull request template:

*
[`.github/pull_request_template.md`](diffhunk://#diff-b2496e80299b8c3150b1944450bd81c622e04e13d15c411d291db0927d75fd6bL3-L4):
Removed the comment suggesting the use of British English for spelling
and terminology.