Update Repository Labels (#24)
## Description

This pull request includes updates to the
`.github/other-configurations/labeller.yml` file to refine the labeling
rules for various types of changes. The most important changes include
adjustments to the documentation and markdown file patterns, as well as
changes to the GitHub Actions file patterns.

Updates to documentation and markdown patterns:

*
[`.github/other-configurations/labeller.yml`](diffhunk://#diff-b490f987d9b055ce6e06e37d13763d61f6e0d5ee349b390bbb910f8d88ff36b9L5-R15):
Modified the `documentation` and `markdown` sections to include
`README.md` and `.github/pull_request_template.md` in the file patterns.

Updates to GitHub Actions patterns:

*
[`.github/other-configurations/labeller.yml`](diffhunk://#diff-b490f987d9b055ce6e06e37d13763d61f6e0d5ee349b390bbb910f8d88ff36b9L20-R26):
Updated the `github_actions` section to include `.github/actions/*` in
the file patterns.