Update GitHub Actions Dependabot Groups (#27)
# Pull Request

## Description

This pull request includes changes to the `.github/dependabot.yml` file
to improve the configuration for maintaining dependencies for GitHub
Actions. The most important changes include renaming a group and
specifying update types for dependencies.

Configuration improvements:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L3):
Removed a comment related to maintaining dependencies for GitHub
Actions.
*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L13-R17):
Renamed the `github-action-dependencies` group to `github-actions` and
specified that only `patch` and `minor` updates should be applied for
dependencies in this group.