Add Code of Conduct (#35)
# Pull Request

## Description

This pull request introduces a new `CODE_OF_CONDUCT.md` file to the
`docs` directory. The document outlines the expected behavior and
community standards for contributors, as well as the enforcement
guidelines and responsibilities for community leaders.

Key additions include:

* Introduction of the Contributor Covenant Code of Conduct, version 2.1,
which sets the tone for a harassment-free and inclusive community.
* Detailed descriptions of acceptable and unacceptable behaviors within
the community.
* Clear enforcement responsibilities for community leaders, including
the right to remove or edit contributions that do not align with the
code of conduct.
* Guidelines for reporting and handling instances of unacceptable
behavior, ensuring privacy and security for reporters.
* Community Impact Guidelines that outline the consequences for
violations, ranging from a written warning to a permanent ban.