Add Security Policy Step (#49)
# Pull Request

## Description

This pull request includes a small change to the `README.md` file to
enhance the setup instructions for a new repository. The change ensures
that the `SECURITY.md` file links to the security tab of the repository.

-
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R25):
Added a step to ensure that the `SECURITY.md` file links to the security
tab of the repository.