Update Dependency Review (#53)
# Pull Request

## Description

This pull request includes changes to the `jobs` section in the
`.github/workflows/pull-request-tasks.yml` file to improve the
dependency review process and its integration with pull requests.

Improvements to dependency review process:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR40-R41):
Added `pull-requests: write` permission to the `dependency-review` job
to enable writing comments on pull requests.
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR50-R51):
Configured the `dependency-review` action to comment a summary in the
pull request only on failure.