Update GitHub Actions (#71)
# Pull Request

## Description

This pull request includes several updates to dependencies and the
removal of unused git hooks. The most important changes include updating
the versions of actions and dependencies in the workflow configuration
files and removing old git hooks scripts.

Updates to dependencies:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L48-R48):
Updated `UmbrellaDocs/action-linkspector` from version `v1.3.2` to
`v1.3.3` to ensure the latest features and fixes are included.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L67-R67):
Updated `extractions/setup-just` from version `v2` to `v3` to use the
latest version of the action.
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Updated the minimum version of `lefthook` from `1.11.2` to `1.11.4` to
ensure compatibility with newer features.

Removal of unused git hooks:

*
[`githooks/post-commit`](diffhunk://#diff-434d9a5655c19f7989a916bcec6e2e60d68a96eae07a53044290623181b520e7L1-L4):
Removed the `post-commit` git hook script as it is no longer needed.
*
[`githooks/pre-commit`](diffhunk://#diff-d5fe9d324948511c959e14d0aa311bb057ce7236626bb86fa51351dc26845a4dL1-L30):
Removed the `pre-commit` git hook script as it is no longer needed.