Remove CodeQL Configuration (#80)
# Pull Request

## Description

This pull request simplifies the CodeQL configuration by removing a
custom configuration file and updating the workflow to use default
settings. The changes focus on improving maintainability and reducing
complexity in the repository's CodeQL setup.

**CodeQL configuration changes:**

*
[`.github/other-configurations/codeql-config.yml`](diffhunk://#diff-687a0af9316072cbdab79cdb046aac181231f8b5f3b5502bc796a9d949ce65beL1-L3):
Removed the custom query filter configuration, which excluded the
`actions/unpinned-tag` query.

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L128):
Updated the CodeQL workflow to stop referencing the removed
configuration file, allowing the default queries to be used instead.