Update Dependabot Schedule (#90)
# Pull Request

## Description

This pull request updates the Dependabot configuration to use a more
flexible scheduling format for GitHub Actions dependency updates.

Configuration changes:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L6-R7):
Changed the `schedule` for the `github-actions` package ecosystem from a
daily interval with a fixed time and timezone to a cron-based schedule
(`30 7 * * *`). This allows for more precise control over when updates
are checked.