Update Dependabot Groups and Configuration (#91)
# Pull Request

## Description

This pull request updates the `.github/dependabot.yml` configuration to
enhance the scheduling and grouping of dependency updates. The most
important changes include specifying a timezone for the cron job and
refining the grouping rules for GitHub Actions updates.

Enhancements to Dependabot configuration:

* Added the `timezone` field to specify "Europe/London" for the cron job
schedule.
* Updated the `github-actions` group to apply specifically to
"version-updates" and exclude patterns matching "super-linter/*".