Remove Pinact Configuration (#94)
# Pull Request

## Description

This pull request removes the use of a custom Pinact configuration file
and updates related commands to use the default configuration. The most
important changes include deleting the custom configuration file and
modifying all references to it in the `Justfile`.

### Removal of custom Pinact configuration:
*
[`.github/other-configurations/pinact.yml`](diffhunk://#diff-c2e684c61036562879b1062fb69ab67d446e19354f7a45ef92b14f611a051bc5L1-L9):
Deleted the custom Pinact configuration file, which is no longer needed.

### Updates to Pinact commands:
*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL55-R63):
Updated all Pinact-related commands (`pinact-run`, `pinact-check`, and
`pinact-update`) to remove the `-c
.github/other-configurations/pinact.yml` option, defaulting to the
built-in configuration instead.