Update Reusable Workflows (#116)
# Pull Request

## Description

This pull request updates the reusable workflow references in multiple
GitHub Actions workflow files to use a newer version (`v2025.05.24.02`).
The updates ensure that the workflows are aligned with the latest
changes and improvements in the reusable workflows.

### Workflow Updates:

* **`.github/workflows/clean-caches.yml`**: Updated the reusable
workflow reference for `common-clean-caches.yml` to the latest version.

* **`.github/workflows/code-checks.yml`**:
- Updated the reusable workflow reference for `common-code-checks.yml`
to the latest version.
- Updated the reusable workflow reference for `codeql-analysis.yml` to
the latest version.

* **`.github/workflows/pull-request-tasks.yml`**: Updated the reusable
workflow reference for `common-pull-request-tasks.yml` to the latest
version.

* **`.github/workflows/sync-labels.yml`**: Updated the reusable workflow
reference for `common-sync-labels.yml` to the latest version.