Add Task to update Readme (#117)
# Pull Request

## Description

This pull request updates the `README.md` file to enhance the setup
instructions for a new repository. The most notable change is the
addition of a step to include project status badges in the `README.md`.

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R22):
Added a new task to include project status badges in the `README.md`
file as part of the repository setup process.