Remove CodeLimit Task (#122)
# Pull Request

## Description

This pull request updates the `README.md` file to revise the list of
tasks for setting up a new repository. The most notable change is the
removal of the step to set up CodeLimit analysis.

Key changes to repository setup tasks:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L56):
Removed the step to set up CodeLimit analysis from the list of tasks for
setting up a new repository.